refactor: consolidate datasource workspace modules
Make lens.datasource the single backend datasource entry point and move session workspace handling into the package. Update internal imports and tests while leaving independent LensNode runtime modules and migration history untouched.
