--- backend/lens/datasource_services.py
@@ -1,7 +1,0 @@
-"""Compatibility import for the consolidated datasource package."""
-
-import importlib
-import sys
-
-_module = importlib.import_module("lens.datasource.services")
-sys.modules[__name__] = _module
