+++ backend/lens/services.py
@@ -64,7 +64,7 @@ from .plugins.registry import installed_plugin
 from .routing_descriptions import build_routing_description
 from .runtime_events import public_step_detail, sanitize_termination_detail
 from .session_lifecycle import lock_active_session
-from .session_workspace import build_session_workspace, session_source_dirs
+from .datasource.workspace import build_session_workspace, session_source_dirs
 from .session_titles import fallback_session_title
 from .trace_context import root_observation_id_for_run, trace_id_for_run
